+++ src/Root.tsx
@@ -12,7 +12,6 @@ import React from "react";
 import { ConfirmationModal } from "./app/Confirmation";
 import { enqueueErrorSnackbar, enqueueSuccessSnackbar, NoticeProvider } from "./app/Notice";
 import { SemVer } from "semver";
-import { useDragDropFile } from "./core/Hook";
 import { APP_VERSION } from "./core/MainApp";
 import { Logger } from "./core/Logger";
 import { checkForUpdates, promptUpdate } from "./core/Versioning";
@@ -24,7 +23,7 @@ import { Diagram } from "./diagram/Diagram";
 import { validate } from "class-validator";
 import { ConfirmationPromptData } from "./core/Confirmation";
 import { onDropFile, onOpen, onSave, onSaveAs, onNew } from "./core/InputOutput";
-import { DragDropBackdrop } from "./app/DragDropBackdrop";
+import { DragDropBackdrop, useDragDropFile } from "./app/DragDropBackdrop";
 
 (window as any)["checkForUpdates"] = checkForUpdates;
 
@@ -106,9 +105,7 @@ export async function handleDiagramParam(encodedDiagramParam: string) {
 
 const Root = observer(() => {
   const { app, modals } = getRootStore();
-  // isUsing is always true for testing PD-68
-  const isUsing = true;
-  const { isDraggingFile, onDragEnter, onDragLeave, onDragOver, onDrop } = useDragDropFile(isUsing, onDropFile);
+
   React.useEffect(() => {
     const searchParams = new URLSearchParams(document.location.search);
     const result = searchParams.get("diagram");
@@ -132,30 +129,9 @@ const Root = observer(() => {
     };
   }, [app]);
 
-  // React.useEffect(() => {
-  // getRootStore().modals.open(HelpModalSymbol);
-  // getRootStore().confirmation.prompt({
-  //   title: "Welcome to Diagrams",
-  //   description: "This is a diagram editor. You can use it to create diagrams.",
-  //   buttons: [
-  //     {
-  //       label: "OK",
-  //       hotkey: "Enter",
-  //       onClick: () => {},
-  //       color: "success"
-  //     },
-  //     {
-  //       label: "Cancel",
-  //       onClick: () => {},
-  //       color: "error"
-  //     }
-  //   ],
-  //   inputLabel: "Name",
-  //   inputDefaultValue: "John",
-  // });
-  // }, []);
-
   const isUsingEditor = !modals.isOpen;
+  const { isDraggingFile, getRootProps: getRootPropsForDragDropBackdrop } = useDragDropFile(isUsingEditor, onDropFile);
+  const { onDragEnter, onDragLeave, onDragOver, onDrop } = getRootPropsForDragDropBackdrop();
 
   const ENABLE_EXCEPT_INPUT_FIELDS = { enabled: isUsingEditor };
 
@@ -183,13 +159,13 @@ const Root = observer(() => {
   useCustomHotkeys("Shift+Mod+Z", handleRedo, ENABLE_EXCEPT_INPUT_FIELDS);
 
   return (
-    <Box id="root-container" {...{ onDragEnter, onDragOver, onDrop }}>
+    <Box id="root-container" {...{onDragEnter, onDragOver, onDrop}}>
       <NoticeProvider />
       <DiagramCanvas />
       <BottomPanel />
       <HelpModal />
       <ConfirmationModal />
-      {isDraggingFile && <DragDropBackdrop {...{ onDragEnter, onDragLeave, onDragOver, onDrop }} />}
+      {isDraggingFile && <DragDropBackdrop {...{onDragEnter, onDragLeave, onDragOver, onDrop}} />}
     </Box>
   );
 });
