+++ src/core/Hook.ts
@@ -145,31 +145,3 @@ export function useCustomHotkeys<T extends HTMLElement>(
     dependencies
   );
 }
-
-export function useDragDropFile(enable: boolean, onDrop: (file: File) => void) {
-  const [isDraggingFile, setIsDraggingFile] = React.useState(false);
-
-  return {
-    isDraggingFile,
-    onDragEnter: (e: React.DragEvent<HTMLDivElement>) => {
-      setIsDraggingFile(e.dataTransfer.types.includes("Files"));
-    },
-    onDragLeave: (e: React.DragEvent<HTMLDivElement>) => {
-      setIsDraggingFile(false);
-    },
-    onDragOver: (e: React.DragEvent<HTMLDivElement>) => {
-      e.preventDefault();
-      e.stopPropagation();
-    },
-    onDrop: (e: React.DragEvent<HTMLDivElement>) => {
-      setIsDraggingFile(false);
-      e.preventDefault();
-      e.stopPropagation();
-      if (enable === false) return;
-
-      const file = e.dataTransfer.files?.[0];
-      if (file === undefined) return;
-      onDrop(file);
-    }
-  };
-}
